Move isExecutionPatchResult to internal
