+++ src/core/QueryManager.ts
@@ -40,7 +40,6 @@ import { maskFragment, maskOperation } from "@apollo/client/masking";
 import type { DeepPartial } from "@apollo/client/utilities";
 import { print } from "@apollo/client/utilities";
 import { AutoCleanedWeakCache, cacheSizes } from "@apollo/client/utilities";
-import { isExecutionPatchResult } from "@apollo/client/utilities";
 import {
   DocumentTransform,
   getGraphQLErrorsFromResult,
@@ -58,6 +57,7 @@ import {
   hasDirectives,
   isDocumentNode,
   isExecutionPatchIncrementalResult,
+  isExecutionPatchResult,
   isNonEmptyArray,
   isNonNullObject,
   onAnyEvent,
